Add single-file x64 publishing profile
Done to make it possible to build an almost self-contained EXE through the "publish" menu in Visual Studio. The EXE still depends on the .Net 6 runtime being installed on the system though.

It's also possible to embed the .Net runtime into the EXE to make it fully self-contained. However, that would increase the EXE size from 0.8MB to 15.3MB (after trimming unused code), which is kind of large for such a tool.