Fix broken IsMachineKey detection on Win11
Fixes #16. The `CngKey.IsMachineKey` property seem to work correctly when testing on Win11 (tested both in VM and on physical HW), whereas it always returns False on Win10.

The GetProperty("Key Type"..) behavior also seem to have changed for LocalMachine key storage, with it returning [1, 0, 0, 0] on Win10 and [32, 0, 0, 0] on Win11.